code review, ipv6 support
